% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_climatebc_locs_normal_1961_1990}
\alias{picola_climatebc_locs_normal_1961_1990}
\title{Site annual climate normal variables 1961-1990}
\format{
A data frame with 18 rows and 30 variables:
\describe{
\item{\code{Site}}{character Short symbolic name of site}
\item{\code{id}}{character The string "site" if this row identifies the exact location of a site, or "grid" if this row identifies the location of a PNWNAmet grid cell enclosing a site}
\item{\code{Latitude}}{double Latitude of site or grid cell}
\item{\code{Longitude}}{double Longitude of site or grid cell}
\item{\code{Elevation}}{integer Elevation of site or grid cell}
\item{\code{MAT}}{double Mean annual temperature (°C)}
\item{\code{MWMT}}{double Mean warmest month temperature (°C)}
\item{\code{MCMT}}{double Mean coldest month temperature (°C)}
\item{\code{TD}}{double Temperature difference between MWMT and MCMT, or continentality (°C)}
\item{\code{MAP}}{integer Mean annual precipitation (mm)}
\item{\code{MSP}}{integer Mean annual summer (May to Sept.) precipitation (mm)}
\item{\code{AHM}}{double Annual heat-moisture index (MAT+10)/(MAP/1000)}
\item{\code{SHM}}{double Summer heat-moisture index ((MWMT)/(MSP/1000)}
\item{\code{DD_0}}{integer Degree-days below 0°C, chilling degree-days}
\item{\code{DD5}}{integer Degree-days above 5°C, growing degree-days}
\item{\code{DD_18}}{integer Degree-days below 18°C, heating degree-days}
\item{\code{DD18}}{integer Degree-days above 18°C, cooling degree-days}
\item{\code{NFFD}}{integer The number of frost-free days}
\item{\code{bFFP}}{integer The day of the year on which frost-free period (FFP) begins}
\item{\code{eFFP}}{integer The day of the year on which frost-free period (FFP) ends}
\item{\code{FFP}}{integer Frost-free period}
\item{\code{PAS}}{integer Precipitation as snow (mm). For individual years, it covers the period between August in the previous year and July in the current year}
\item{\code{EMT}}{double Extreme minimum temperature over 30 years (°C)}
\item{\code{EXT}}{double Extreme maximum temperature over 30 years (°C)}
\item{\code{MAR}}{double Mean annual solar radiation (MJ m‐2 d‐1)}
\item{\code{Eref}}{integer Hargreaves reference evaporation (mm)}
\item{\code{CMD}}{integer Hargreaves climatic moisture deficit (mm)}
\item{\code{RH}}{integer Mean annual relative humidity (\%)}
\item{\code{CMI}}{double Hogg’s climate moisture index (mm)}
\item{\code{DD1040}}{integer Degree-days above 10°C and below 40°C}
}
}
\source{
ClimateBC v7.30 https://climatebc.ca/
}
\usage{
picola_climatebc_locs_normal_1961_1990
}
\description{
Annual climate variables for normal period 1961-1990 for British Columbia Pinus contorta ssp. latifolia seed orchard and northern comparison sites in British Columbia, as well as enclosing PNWNAmet grid cells
}
\details{
This dataset provides annual climate normal variables for the period 1961-1990 for seed orchard sites located in British Columbia, as well as for the enclosing PNWNAmet grid cells. These climate variables are relevant for understanding the climatic conditions in the seed orchards.

The dataset includes both the exact locations of seed orchard sites and the centers of the PNWNAmet grid cells that enclose these sites. Each row in the dataset corresponds to either a seed orchard site or a grid cell center.

The climate data were derived from ClimateBC v7.30, which provides spatially downscaled and customizable climate data for historical and future periods.
}
\references{
Wang T, Hamann A, Spittlehouse D, Carroll C (2016) Locally Downscaled and Spatially Customizable Climate Data for Historical and Future Periods for North America. PLoS ONE 11(6): e0156720. doi:10.1371/journal.pone.0156720
}
\keyword{datasets}
