% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_climatebc_locs_normal_monthly}
\alias{picola_climatebc_locs_normal_monthly}
\title{Site monthly climate normal variables 1921-2020}
\format{
A data frame with 108 rows and 66 variables:
\describe{
\item{\code{Site}}{character Short symbolic name of site}
\item{\code{id}}{character The string "site" if this row identifies the exact location of a site, or "grid" if this row identifies the location of a PNWNAmet grid cell enclosing a site}
\item{\code{Latitude}}{double Latitude of site or grid cell}
\item{\code{Longitude}}{double Longitude of site or grid cell}
\item{\code{Elevation}}{integer Elevation of site or grid cell}
\item{\code{Tmax01}}{double January maximum mean temperature (°C)}
\item{\code{Tmax02}}{double February maximum mean temperature (°C)}
\item{\code{Tmax03}}{double March maximum mean temperature (°C)}
\item{\code{Tmax04}}{double April maximum mean temperature (°C)}
\item{\code{Tmax05}}{double May maximum mean temperature (°C)}
\item{\code{Tmax06}}{double June maximum mean temperature (°C)}
\item{\code{Tmax07}}{double July maximum mean temperature (°C)}
\item{\code{Tmax08}}{double August maximum mean temperature (°C)}
\item{\code{Tmax09}}{double September maximum mean temperature (°C)}
\item{\code{Tmax10}}{double October maximum mean temperature (°C)}
\item{\code{Tmax11}}{double November maximum mean temperature (°C)}
\item{\code{Tmax12}}{double December maximum mean temperature (°C)}
\item{\code{Tmin01}}{double January minimum mean temperature (°C)}
\item{\code{Tmin02}}{double February minimum mean temperature (°C)}
\item{\code{Tmin03}}{double March minimum mean temperature (°C)}
\item{\code{Tmin04}}{double April minimum mean temperature (°C)}
\item{\code{Tmin05}}{double May minimum mean temperature (°C)}
\item{\code{Tmin06}}{double June minimum mean temperature (°C)}
\item{\code{Tmin07}}{double July minimum mean temperature (°C)}
\item{\code{Tmin08}}{double August minimum mean temperature (°C)}
\item{\code{Tmin09}}{double September minimum mean temperature (°C)}
\item{\code{Tmin10}}{double October minimum mean temperature (°C)}
\item{\code{Tmin11}}{double November minimum mean temperature (°C)}
\item{\code{Tmin12}}{double December minimum mean temperature (°C)}
\item{\code{Tave01}}{double January mean temperature (°C)}
\item{\code{Tave02}}{double February mean temperature (°C)}
\item{\code{Tave03}}{double March mean temperature (°C)}
\item{\code{Tave04}}{double April mean temperature (°C)}
\item{\code{Tave05}}{double May mean temperature (°C)}
\item{\code{Tave06}}{double June mean temperature (°C)}
\item{\code{Tave07}}{double July mean temperature (°C)}
\item{\code{Tave08}}{double August mean temperature (°C)}
\item{\code{Tave09}}{double September mean temperature (°C)}
\item{\code{Tave10}}{double October mean temperature (°C)}
\item{\code{Tave11}}{double November mean temperature (°C)}
\item{\code{Tave12}}{double December mean temperature (°C)}
\item{\code{PPT01}}{integer January precipitation (mm)}
\item{\code{PPT02}}{integer February precipitation (mm)}
\item{\code{PPT03}}{integer March precipitation (mm)}
\item{\code{PPT04}}{integer April precipitation (mm)}
\item{\code{PPT05}}{integer May precipitation (mm)}
\item{\code{PPT06}}{integer June precipitation (mm)}
\item{\code{PPT07}}{integer July precipitation (mm)}
\item{\code{PPT08}}{integer August precipitation (mm)}
\item{\code{PPT09}}{integer September precipitation (mm)}
\item{\code{PPT10}}{integer October precipitation (mm)}
\item{\code{PPT11}}{integer November precipitation (mm)}
\item{\code{PPT12}}{integer December precipitation (mm)}
\item{\code{Rad01}}{double January solar radiation (MJ m-2 d-1)}
\item{\code{Rad02}}{double February solar radiation (MJ m-2 d-1)}
\item{\code{Rad03}}{double March solar radiation (MJ m-2 d-1)}
\item{\code{Rad04}}{double April solar radiation (MJ m-2 d-1)}
\item{\code{Rad05}}{double May solar radiation (MJ m-2 d-1)}
\item{\code{Rad06}}{double June solar radiation (MJ m-2 d-1)}
\item{\code{Rad07}}{double July solar radiation (MJ m-2 d-1)}
\item{\code{Rad08}}{double August solar radiation (MJ m-2 d-1)}
\item{\code{Rad09}}{double September solar radiation (MJ m-2 d-1)}
\item{\code{Rad10}}{double October solar radiation (MJ m-2 d-1)}
\item{\code{Rad11}}{double November solar radiation (MJ m-2 d-1)}
\item{\code{Rad12}}{double December solar radiation (MJ m-2 d-1)}
\item{\code{normal_period}}{character 30-year normal period}
}
}
\source{
ClimateBC v7.30 https://climatebc.ca/
}
\usage{
picola_climatebc_locs_normal_monthly
}
\description{
Monthly climate variables for 30-year periods from 1921-1950 for \emph{Pinus contorta} ssp. \emph{latifolia} seed orchard and northern comparison sites in British Columbia, as well as enclosing PNWNAmet grid cells
}
\details{
This dataset provides monthly climate normal variables for the period 1921-2020 for \emph{Pinus contorta} ssp. \emph{latifolia} seed orchard and comparison sites in British Columbia, as well as for the enclosing PNWNAmet grid cells. The climate data include variables such as temperature, precipitation, and solar radiation, all aggregated into 30-year periods every 10 years starting in 1921.

The dataset includes both the exact locations of sites and the centers of the PNWNAmet grid cells that enclose these sites. Each row in the dataset corresponds to either a site or a grid cell center near that site.

The data were derived from ClimateBC v7.30, which provides downscaled and customizable climate data for both historical and future periods.
}
\references{
Wang T, Hamann A, Spittlehouse D, Carroll C (2016) Locally Downscaled and Spatially Customizable Climate Data for Historical and Future Periods for North America. PLoS ONE 11(6): e0156720. doi:10.1371/journal.pone.0156720
}
\keyword{datasets}
