% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_normal_daily}
\alias{picola_normal_daily}
\title{Normal period daily temperatures 1921-2100}
\format{
A data frame with 59292 rows and 7 variables:
\describe{
\item{\code{Site}}{character Short symbolic name of site}
\item{\code{period}}{character 30 year period for which the data are provided}
\item{\code{scenario}}{character The string "historical" if the data are for a historical period, or "ssp126", "ssp245", "ssp370", "ssp585" if the data are for a future IPCC Shared Socioeconomic Pathway (SSP) scenario}
\item{\code{month}}{integer Month number (1-12)}
\item{\code{DoY}}{integer Day of the year (1-365)}
\item{\code{Date_scale}}{character Representative datestamp for the day of the year (with dummy year 2024)}
\item{\code{mean_temp}}{double Mean temperature (°C)}
}
}
\source{
ClimateBC v7.30 https://climatebc.ca/
}
\usage{
picola_normal_daily
}
\description{
Estimated daily temperature data for 30-year normal periods at seed orchard and comparison sites from 1921-2100.
}
\details{
This dataset provides estimated daily temperature data for 30-year normal periods at \emph{Pinus contorta} ssp. \emph{latifolia} seed orchard and comparison sites in British Columbia, ranging from 1921 to 2100. The data combines historical daily temperatures and future projections under different IPCC Shared Socioeconomic Pathway (SSP) scenarios: SSP126, SSP245, SSP370, and SSP585, which correspond to different levels of radiative forcing (from 4.5 to 8.5 W/m²) by 2100.

Monthly climate normals for historical and future periods are obtained from ClimateBC v7.30. These climate normals are used to adjust a "typical" year of the 1941-2012 adjusted PWNAmet daily temperatures (\strong{\code{picola_PNWNAmet_adjusted}}) to match the climate normals for each period. The typical year is generated by averaging the daily temperatures across the full range of historical data (1945-2012), and the monthly climate normals from ClimateBC are applied to either increase or decrease the daily temperatures to align with the projected or historical monthly mean temperatures.
}
\references{
Wang T, Hamann A, Spittlehouse D, Carroll C (2016) Locally Downscaled and Spatially Customizable Climate Data for Historical and Future Periods for North America. PLoS ONE 11(6): e0156720. doi:10.1371/journal.pone.0156720
Mahony, CR, Wang, T; Hamann, A and Cannon, AJ, 2022. A CMIP6 ensemble for downscaled monthly climate normals over North America. International Journal of Climatology 42 (11), 5871-5891 DOI: https://doi.org/10.1002/joc.7566
}
\keyword{datasets}
