% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_PNWNAmet_adjusted}
\alias{picola_PNWNAmet_adjusted}
\title{PNWNAmet daily temperature 1945-2012}
\format{
A data frame with 223533 rows and 8 variables:
\describe{
\item{\code{Date}}{character Date of the temperature measurements}
\item{\code{Site}}{character Short symbolic name of site}
\item{\code{max_temp}}{double Maximum temperature (째C)}
\item{\code{min_temp}}{double Minimum temperature (째C)}
\item{\code{mean_temp_raw}}{double Mean temperature (째C) as measured by PNWNAmet}
\item{\code{intercept}}{double Intercept of the adjustment equation}
\item{\code{slope}}{double Slope of the adjustment equation}
\item{\code{mean_temp_corrected}}{double Mean temperature (째C) adjusted by monthly temperatures from ClimateBC v7.30}
}
}
\source{
Pacific Climate Impacts Consortium https://www.pacificclimate.org/ and ClimateBC v7.30 https://climatebc.ca/
}
\usage{
picola_PNWNAmet_adjusted
}
\description{
PNWNAmet daily temperature data adjusted by monthly temperatures from ClimateBC v7.30 for the period 1945-2012 for \emph{Pinus contorta} ssp. \emph{latifolia} seed orchard and comparison sites in British Columbia
}
\details{
This dataset contains daily temperature data for \emph{Pinus contorta} ssp. \emph{latifolia} seed orchard and comparison sites in British Columbia from 1945-2012. Data was obtained from the Pacific Climate Impacts Consortium PNWNAmet dataset and then adjusted by monthly temperature data from ClimateBC v7.30. The data includes daily maximum, minimum, and mean temperatures, as well as correction factors (intercept and slope) derived from the regression of PNWNAmet data on monthly ClimateBC data. These adjustments were necessary due to the offsets between PNWNAmet grid cell centers and site locations.

The adjustment process involved regressing monthly mean temperatures from PNWNAmet against corresponding monthly temperatures from ClimateBC v7.30. The intercept and slope from this regression were used to adjust the daily temperatures from PNWNAmet, ensuring that they better matched the ClimateBC projections for the given time period.
For details on the methodology for calculating these adjustment factors, refer to the \code{picola_corrframemo} dataset and its accompanying documentation.
}
\references{
Werner, A.T., Schnorbus, M.A., Shrestha, R.R., Cannon, A.J., Zwiers, F.W., Dayon G. and Anslow, F., 2019. A long-term, temporally consistent, gridded daily meteorological dataset for northwestern North America, Scientific Data, 6, 180299, doi:10.1038/sdata.2018.299
Wang T, Hamann A, Spittlehouse D, Carroll C (2016) Locally Downscaled and Spatially Customizable Climate Data for Historical and Future Periods for North America. PLoS ONE 11(6): e0156720. doi:10.1371/journal.pone.0156720
}
\keyword{datasets}
