% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_SPU_climsum}
\alias{picola_SPU_climsum}
\title{SPU Annual Climate Normals}
\format{
A data frame with 32 rows and 26 variables:
\describe{
\item{\code{SPU_Number}}{integer Seed planning unit numerical code}
\item{\code{Pl_SPU}}{character Seed planning unit name}
\item{\code{X_FREQ_}}{integer Unknown. Believed to have something to do with the size or shape of the SPU.}
\item{\code{Latitude}}{double Latitude of the SPU}
\item{\code{Longitude}}{double Longitude of the SPU}
\item{\code{Elevation}}{integer Elevations of the SPU (m)}
\item{\code{MAT}}{double Mean annual temperature (°C)}
\item{\code{MWMT}}{double Mean warmest month temperature (°C)}
\item{\code{MCMT}}{double Mean coldest month temperature (°C)}
\item{\code{TD}}{double Temperature difference between MWMT and MCMT, or continentality (°C)}
\item{\code{MAP}}{integer Mean annual precipitation (mm)}
\item{\code{MSP}}{integer Mean annual summer (May to Sept.) precipitation (mm)}
\item{\code{AHM}}{double Annual heat-moisture index (MAT+10)/(MAP/1000)}
\item{\code{SHM}}{double Summer heat-moisture index ((MWMT)/(MSP/1000)}
\item{\code{DD_0}}{integer Degree-days below 0°C, chilling degree-days}
\item{\code{DD5}}{integer Degree-days above 5°C, growing degree-days}
\item{\code{DD_18}}{integer Degree-days below 18°C, heating degree-days}
\item{\code{DD18}}{integer Degree-days above 18°C, cooling degree-days}
\item{\code{NFFD}}{integer The number of frost-free days}
\item{\code{bFFP}}{integer The day of the year on which frost-free period (FFP) begins}
\item{\code{eFFP}}{integer The day of the year on which frost-free period (FFP) ends}
\item{\code{FFP}}{integer Frost-free period}
\item{\code{PAS}}{integer Precipitation as snow (mm). For individual years, it covers the period between August in the previous year and July in the current year}
\item{\code{EMT}}{double Extreme minimum temperature over 30 years (°C)}
\item{\code{Eref}}{integer Hargreaves reference evaporation (mm)}
\item{\code{CMD}}{integer Hargreaves climatic moisture deficit (mm)}
}
#' @details
This dataset contains climate normals for \emph{Pinus contorta} ssp. \emph{latifolia} Seed Planning Units (SPUs) in British Columbia. The SPUs are geographic units used for organizing breeding and seed production activities based on species, seed planning zone, and elevation band. These 30-year climate normals are derived from ClimateBC, which provides high-resolution downscaled climate data for the region.

The data cover key climatic indicators, including the mean annual temperature (MAT), mean warmest month temperature (MWMT), precipitation levels, and heat-moisture indices, which collectively characterize the climate conditions across each SPU.
}
\source{
ClimateBC v7.30 https://climatebc.ca/
}
\usage{
picola_SPU_climsum
}
\description{
Climate normals for BC \emph{Pinus contorta} ssp. \emph{latifolia} Seed Planning Units.
}
\references{
Wang T, Hamann A, Spittlehouse D, Carroll C (2016) Locally Downscaled and Spatially Customizable Climate Data for Historical and Future Periods for North America. PLoS ONE 11(6): e0156720. doi:10.1371/journal.pone.0156720
}
\keyword{datasets}
