% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{picola_typical_ts}
\alias{picola_typical_ts}
\title{Typical year daily temperatures}
\format{
A data frame with 3294 rows and 5 variables:
\describe{
\item{\code{Site}}{character Short symbolic name of site}
\item{\code{DoY}}{integer Day of the year (1-365)}
\item{\code{mean_mean_temp}}{double mean of mean daily temperature 1945-2012}
\item{\code{Date_scale}}{character Representative datestamp for the day of the year (in 2024)}
\item{\code{month}}{integer Month number (1-12)}
}
}
\usage{
picola_typical_ts
}
\description{
Daily mean temperatures for a "typical" year across various seed orchard sites and two northern comparison sites in British Columbia.
}
\details{
The "typical year" daily temperature values were derived by averaging the mean daily temperatures across all years from 1945 to 2012, which represents the full temporal range of the PNWNAmet dataset. This method provides a general representation of temperature patterns at each site, smoothing out annual variation and offering a reference for climate conditions during a typical year.

The raw PNWNAmet data was adjusted using the correction factors derived from ClimateBC data. This correction process aligns the PNWNAmet data with the more accurate monthly temperature data from ClimateBC, ensuring that the daily temperatures in this dataset reflect a more accurate representation of typical climate conditions for each site.
}
\keyword{datasets}
